src/requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
argcomplete==3.5.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==24.2.0
boto3==1.35.9
botocore==1.35.9
certifi==2024.8.30
cffi==1.16.0
cfn-flip==1.3.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==42.0.7
deprecation==2.1.0
dj-database-url==2.2.0
Django==4.2.15
django-cors-headers==4.4.0
django-rest-knox==5.0.1
django-storages==1.14.4
djangorestframework==3.15.2
durationpy==0.7
exceptiongroup==1.2.2
factory_boy==3.3.1
Faker==28.0.0
frozenlist==1.4.1
gotrue==2.7.0
gunicorn==23.0.0
h11==0.14.0
h2==4.1.0
hjson==3.1.0
hpack==4.0.0
httpcore==1.0.5
httpx==0.27.2
hyperframe==6.0.1
idna==3.8
jmespath==1.0.1
kappa==0.6.0
MarkupSafe==2.1.5
multidict==6.0.5
packaging==24.1
placebo==0.9.0
postgrest==0.16.11
psycopg==3.1.10
psycopg-binary==3.1.10
psycopg2==2.9.9
psycopg2-binary==2.9.7
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
python-dateutil==2.9.0.post0
python-slugify==8.0.4
PyYAML==6.0.2
realtime==2.0.2
requests==2.32.3
ruff==0.6.3
s3transfer==0.10.2
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.1
storage3==0.7.7
StrEnum==0.4.15
supabase==2.7.3
supafunc==0.5.1
text-unidecode==1.3
toml==0.10.2
tqdm==4.66.5
troposphere==4.8.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==1.26.20
websockets==12.0
Werkzeug==3.0.4
whitenoise==6.7.0
yarl==1.9.4
zappa==0.59.0
